create dir for config
