+++ lib/config/index.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const path = require('path');
-const v = require('@mapbox/fusspot');
 const _ = require('lodash');
-const { WEBPACK_ASSETS_BASENAME } = require('./constants');
+const { WEBPACK_ASSETS_BASENAME } = require('../constants');
 const fs = require('fs');
+const validateConfig = require('./validate-config');
 
 module.exports = createUrc;
 
@@ -48,47 +48,4 @@ class Urc {
 
 function createUrc(rawConfig = {}, defaultConfig) {
   return new Urc(rawConfig, defaultConfig);
-}
-
-function validateConfig(rawConfig = {}) {
-  try {
-    v.assert(
-      v.strictShape({
-        siteBasePath: v.string,
-        stylesheets: v.arrayOf(v.string),
-        vendorModules: v.arrayOf(v.string),
-        environmentVariables: v.arrayOf(v.string),
-        postcssPlugins: v.arrayOf(v.func),
-        port: v.number,
-        publicAssetsPath: v.string,
-        devServerHistoryFallback: v.boolean,
-        babelPlugins: v.arrayOf(validateAbsolutePaths),
-        babelPresets: v.arrayOf(validateAbsolutePaths),
-        webpackLoaders: v.arrayOf(v.plainObject),
-        webpackPlugins: v.arrayOf(v.any),
-        webpackConfigTransform: v.func,
-        outputDirectory: v.string,
-        publicDirectory: v.string,
-        jsEntry: v.string,
-        htmlSource: v.string,
-        polyfills: v.strictShape({
-          objectAssign: v.boolean,
-          promise: v.boolean,
-          fetch: v.boolean
-        }),
-        production: v.boolean,
-        stats: v.string
-      })
-    )(rawConfig);
-  } catch (error) {
-    error.message = `Invalid Underreact configuration. ${error.message}`;
-    throw error;
-  }
-}
-
-// Custom fusspot validator.
-function validateAbsolutePaths(value) {
-  if (typeof value !== 'string' || !path.isAbsolute(value)) {
-    return 'absolute path';
-  }
 }
