+++ lib/config/index.test.js
@@ -1,8 +1,9 @@
 'use strict';
 
-const normalizeConfig = require('./normalize-config');
 const path = require('path');
-const getDefaultConfig = require('./default-underreact.config');
+
+const getDefaultConfig = require('../default-underreact.config');
+const createConfig = require('./');
 
 test('normalizes empty configuration', () => {
   const defaultConfig = getDefaultConfig({
@@ -11,7 +12,7 @@ test('normalizes empty configuration', () => {
     port: undefined,
     configDir: path.join(process.cwd(), 'underreact.config.js')
   });
-  expect(normalizeConfig({}, defaultConfig)).toMatchSnapshot();
+  expect(createConfig({}, defaultConfig)).toMatchSnapshot();
 });
 
 test('normalizes fancy configuration', () => {
@@ -22,7 +23,7 @@ test('normalizes fancy configuration', () => {
     configDir: path.join(process.cwd(), 'underreact.config.js')
   });
   expect(
-    normalizeConfig(
+    createConfig(
       {
         polyfills: {
           promise: true,
